+++ src/app/d2-loadout-builder/locked-armor/SelectableBungieImage.m.scss
@@ -36,12 +36,12 @@
 }
 
 .goodPerk {
-  background-image: url('../../../../images/good.png');
+  background-image: url('../../../images/good.png');
   background-size: 10px 10px;
 }
 
 .badPerk {
-  background-image: url('../../../../images/bad.png');
+  background-image: url('../../../images/bad.png');
   background-size: 10px 10px;
 }
