+++ src/app/d2-loadout-builder/locked-armor/SelectableBungieImage.m.scss.d.ts
