--- src/app/d2-loadout-builder/locked-armor/lockeditemcontainer.scss
@@ -1,29 +1,0 @@
-.locked-items {
-  flex: 1;
-}
-.locked-item {
-  text-align: center;
-  position: relative;
-  margin-bottom: calc(var(--item-margin) / 2);
-
-  > div:first-child {
-    display: flex;
-    flex-wrap: wrap;
-  }
-  .locked-item-box {
-    position: relative;
-  }
-  .empty-item {
-    cursor: pointer;
-    // width: calc(var(--item-size) + 4px);
-    // height: calc(var(--item-size) + 4px);
-    width: var(--item-size);
-    height: var(--item-size);
-    margin: 0 calc(var(--item-margin) / 2) calc(var(--item-margin) / 2) 0;
-    position: relative;
-    contain: strict;
-    box-sizing: border-box;
-    transition: opacity 0.2s, transform 0.2s;
-    border: 1px solid #ddd;
-  }
-}
