--- src/app/d2-loadout-builder/locked-armor/popup/SelectableBurn.tsx
@@ -1,34 +1,0 @@
-import classNames from 'classnames';
-import React from 'react';
-import { LockedItemType, BurnItem } from '../../types';
-
-export default function SelectableBurn({
-  burn,
-  selected,
-  onLockedPerk,
-  onHoveredPerk
-}: {
-  burn: BurnItem;
-  selected: boolean;
-  onLockedPerk(burn: LockedItemType): void;
-  onHoveredPerk(burn: {}): void;
-}) {
-  const handleClick = () => {
-    onLockedPerk({ type: 'burn', item: burn });
-  };
-  const handleHover = () => {
-    onHoveredPerk(burn.displayProperties);
-  };
-
-  return (
-    <img
-      className={classNames(`perk-image ${burn.index}`, {
-        'locked-perk': selected
-      })}
-      title={burn.displayProperties.name}
-      onClick={handleClick}
-      onMouseEnter={handleHover}
-      src={burn.displayProperties.icon}
-    />
-  );
-}
